requirements.txt
flask
flask-cors
numpy
pillow
werkzeug
gunicorn
tflite-runtime
